Improve performance for block loading Fix some bossbar things (showing bossbar even though the player can't see it) Change log4j2 into info mode

+++ gomint-server/src/main/java/io/gomint/server/world/ChunkSlice.java
@@ -100,7 +100,7 @@ class ChunkSlice {
     }
 
     void setBlock( int x, int y, int z, byte blockId ) {
-        int index = getIndex( x, y, z );
+        short index = getIndex( x, y, z );
 
         if ( blockId != 0 ) {
             if ( this.blocks == null ) {
