+++ gomint-server/src/main/java/io/gomint/server/world/block/Beetroot.java
@@ -18,8 +18,8 @@ import java.util.List;
 public class Beetroot extends Growable implements BlockBeetroot {
 
     @Override
-    public int getBlockId() {
-        return 244;
+    public byte getBlockId() {
+        return (byte) 244;
     }
 
     @Override
