+++ gomint-server/src/main/java/io/gomint/server/world/block/BrownGlazedTerracotta.java
@@ -12,8 +12,8 @@ import io.gomint.server.registry.RegisterInfo;
 public class BrownGlazedTerracotta extends Block implements io.gomint.world.block.BlockBrownGlazedTerracotta {
 
     @Override
-    public int getBlockId() {
-        return 232;
+    public byte getBlockId() {
+        return (byte) 232;
     }
 
     @Override
