+++ gomint-server/src/main/java/io/gomint/server/world/block/Concrete.java
@@ -12,14 +12,15 @@ import io.gomint.server.registry.RegisterInfo;
 public class Concrete extends Block implements io.gomint.world.block.BlockConcrete {
 
     @Override
-    public int getBlockId() {
-        return 236;
+    public byte getBlockId() {
+        return (byte) 236;
     }
 
     @Override
     public long getBreakTime() {
         return 2700;
     }
+
     @Override
     public float getBlastResistance() {
         return 9.0f;
