+++ gomint-server/src/main/java/io/gomint/server/world/block/Flower.java
@@ -12,7 +12,7 @@ import io.gomint.server.registry.RegisterInfo;
 public class Flower extends Block implements io.gomint.world.block.BlockFlower {
 
     @Override
-    public int getBlockId() {
+    public byte getBlockId() {
         return 38;
     }
