+++ gomint-server/src/main/java/io/gomint/server/world/block/Podzol.java
@@ -12,8 +12,8 @@ import io.gomint.server.registry.RegisterInfo;
 public class Podzol extends Block implements io.gomint.world.block.BlockPodzol {
 
     @Override
-    public int getBlockId() {
-        return 243;
+    public byte getBlockId() {
+        return (byte) 243;
     }
 
     @Override
